Move parallel tutorial example into heartbeat
